+++ blog-admin/src/main/java/com/blog/mapper/BlogTagMapper.java
@@ -5,55 +5,22 @@ import org.apache.ibatis.annotations.Param;
 
 import java.util.List;
 
-/**
- * @author: zhang
- * @time: 2024-09-14 11:17
- */
 public interface BlogTagMapper {
-    /**
-     * 根据主键删除
-     * @param id 博客标签关联id
-     * @return int
-     */
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
-    /**
-     * 插入博客标签关联信息
-     * @param record 博客标签关联信息
-     * @return int
-     */
     int insert(BlogTag record);
 
-    /**
-     * 选择性地插入博客标签关联信息
-     * @param record 博客标签关联信息
-     * @return int
-     */
     int insertSelective(BlogTag record);
 
-    /**
-     * 根据主键查询博客标签关联信息
-     * @param id 博客标签关联id
-     * @return BLogTag
-     */
-    BlogTag selectByPrimaryKey(Integer id);
+    BlogTag selectByPrimaryKey(Long id);
 
-    /**
-     * 根据主键选择性地更新博客标签关联信息
-     * @param record 博客标签关联信息
-     * @return int
-     */
     int updateByPrimaryKeySelective(BlogTag record);
 
-    /**
-     * 根据主键更新博客标签关联信息
-     * @param record 博客标签关联信息
-     * @return int
-     */
     int updateByPrimaryKey(BlogTag record);
 
     /**
      * 批量插入博客标签关联信息
+     *
      * @param blogTags 博客标签关联信息
      * @return int
      */
@@ -61,6 +28,7 @@ public interface BlogTagMapper {
 
     /**
      * 根据博客id查询博客标签关联信息
+     *
      * @param blogId 博客id
      * @return List<BlogTag>
      */
@@ -68,6 +36,7 @@ public interface BlogTagMapper {
 
     /**
      * 根据博客id删除博客标签关联信息
+     *
      * @param blogId 博客id
      */
     void deleteByBlogId(Long blogId);
